vfe/requirements.txt
audioread==2.1.6
cffi==1.12.2
cycler==0.10.0
decorator==4.4.0
editdistance==0.5.3
imageio==2.5.0
joblib==0.13.2
kiwisolver==1.0.1
librosa==0.6.3
llvmlite==0.28.0
matplotlib==3.0.3
networkx==2.3
numba==0.43.1
numpy==1.16.2
opencv-python==4.0.1.23
Pillow==6.0.0
pkg-resources==0.0.0
pycparser==2.19
pyparsing==2.4.0
python-dateutil==2.8.0
python-speech-features==0.6
PyWavelets==1.0.3
resampy==0.2.1
scikit-image==0.15.0
scikit-learn==0.20.3
scikit-video==1.1.11
scipy==1.2.1
six==1.12.0
SoundFile==0.10.2
torch==1.1.0
torchvision==0.3.0
